Move renderScoreBox to templates/misc.dart
